feat: hashing and caching
New command-line option `--cache-to`: if set, the target is used as a cache in terms of:
- the set of explicitly provided actual source files
- the hashes of all processed files (including those loaded through search paths)

If either change, the pruning is re-run, if not, the result of the previous invocation is printed as-is. The aim is to avoid costly re-elaboration if possible.

Two new libraries were integrated for this purpose, nlohmann::json and picosha2.